requirements.txt
typer[all]>=0.9.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
# hashlib is a built-in module
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
openai>=1.0.0
rapidfuzz>=3.0.0
pgvector>=0.2.0